Nuke a bunch of unused code.
I nuked all this code because it is currently unused. I know that it is a
start in the direction that we want to move. However, it was done in a middle
out manner instead of an outside in manner. Therefore, we must nuke it and
start over with actual BDD outside in development.