Made shanty init require language.
I did this so that we could build a working .shanty.yml file for the user via
the shanty init command. I also reworked the command so that it always tries
to generate the .shanty.yml and allows Thor to prompt the user to overwrite,
diff, etc when it already exists.