Add virtualbox provider.
I did this so that I could easily test the entire shanty code base and process
locally without having to be connected to a cloud provider. It also gives
users a simpler path to test their setup with before switching over to a
cloud provider so that the only thing they should have to worry about is
cloud configuration details ideally.