Bubble up test runner exit code and out shanty.
I added support for getting the exit code of the test runner that is remotely
run and bubbling it back up through `shanty test` so that if a CI system is
running shanty or a user for that matter. They can determine of the entire
test suite was successful or if there was a failure. Also, I managed to get
the test runners stdout and stderr to output as the data comes back over ssh
rather than waiting until the end and dumping it all at once.