Fix test runner to eliminate multiple script calls.
I had one of the helper methods still calling script hard coded rather than
using the passed in hook name. This was causing it to duplicate the script
commands in the test runner. This corrects this bug which elminates the
duplication of the script commands and fixes the other hooks that would have
previously been ignored.