Add SIGTERM handler to run cleanup in test.
I did this because some CI systems, Bamboo for one, allow you to stop a build
which send a SIGTERM to the bash script that runs Shanty, which the bash
script passes onto Shanty if the bash script itself doesn't explicitly handle
it. Plus a user might also want to stop the process locally on their machine.