Add comments so I don't forget why I did stuff.
I added comments around the specific sections of code where I had to do
special stuff to handle Bamboo's Stop Build feature because it is very
specialized and a little weird. Also, it is not the correct way long term to
solve the problem. The code that is there now was just a spike so that I could
verify exactly what the problem was and come up with a temporary hack solution
to get things working for teams ASAP. The correct way to solve this problem is
to create an Output layer and a Command Runner layer in the code base and have
each of those have the specific logic to handle the Errno::EPIPE scenarios
properly.